test(arel): consolidate 1-excess duplicate describe tail in attribute.test.ts
