+++ packages/arel/src/attributes/attribute.trails.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { testConnection } from "../test-helpers/connection.js";
-import { Table, Nodes, Visitors } from "../index.js";
+import { Table, Nodes, Visitors, relationName } from "../index.js";
 
 // TS-only coverage for the `when Enumerable` arm of Attribute#in / #notIn
 // (arel/predications.rb:65-74, 112-121). Ruby's Enumerable spans Set, Hash and
@@ -87,6 +87,27 @@ describe("AttributeTest (trails)", () => {
       expect(attribute.notIn(map)).toEqual(
         new Nodes.NotIn(attribute, [new Nodes.Casted(["a", 1], attribute)]),
       );
+    });
+  });
+
+  describe("relationName", () => {
+    it("returns a plain string name unchanged", () => {
+      expect(relationName("users")).toBe("users");
+    });
+
+    it("unwraps a SqlLiteral name to its value", () => {
+      // Unlike Rails (where SqlLiteral < String), trails models SqlLiteral as a
+      // standalone Node, so a TableAlias derived-table alias must be unwrapped
+      // before it flows into any String-typed slot.
+      const alias = new Nodes.TableAlias(new Table("users"), new Nodes.SqlLiteral("derived"));
+      expect(relationName(alias.name)).toBe("derived");
+    });
+
+    it("coerces an attribute on a SqlLiteral-aliased relation to a string name", () => {
+      const alias = new Nodes.TableAlias(new Table("users"), new Nodes.SqlLiteral("derived"));
+      const attr = alias.get("id");
+      const qualified = `${relationName(attr.relation.name)}.${attr.name}`;
+      expect(qualified).toBe("derived.id");
     });
   });
 });
